debug

+++ src/ml/train/finetune.py
@@ -47,7 +47,7 @@ class Finetune:
             load_in_4bit=True,
             bnb_4bit_use_double_quant=True,
             bnb_4bit_quant_type="nf4",
-            bnb_4bit_compute_dtype=torch.bfloat16
+            bnb_4bit_compute_dtype=torch.float16
         )
         model = AutoModelForCausalLM.from_pretrained(self.model_config['base_model_id'],
                                                      quantization_config=bnb_config)
